// src/models/HttpErrors.ts
export enum HttpError {
  badRequest = 'badRequest',
  unauthorized = 'unauthorized',
  forbidden = 'forbidden',
  notFound = 'notFound',
  conflitc = 'conflitc',
  serverError = 'serverError',
  invalidData = 'invalidData',
  unprocessableEntity = 'unprocessableEntity',
  exceedsFileSize = 'exceedsFileSize',
  invalidFormat = 'invalidFormat',
  geolocationError = 'geolocationError',
}